refactor CaseActions to simplify testplans.views.cases view method
Transaction is involved to ensure various case actions are atomic.
